requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu130

# Core scientific stack
numpy==1.26.4
pandas==2.3.1
scipy==1.16.1
scikit-learn==1.7.1
matplotlib==3.10.5
Pillow==11.3.0
PyYAML==6.0.2
requests==2.32.4
tqdm==4.67.1
joblib==1.5.1
python-dateutil==2.9.0.post0
attrs==25.3.0
tenacity==9.1.2
python-dotenv==1.1.1

# Deep learning (CUDA build pulled from the extra index)
tensorflow==2.17.0
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
ultralytics==8.3.203

# Computer vision
opencv-python==4.11.0.86
opencv-python-headless>=4.9
opencv-contrib-python==4.11.0.86
mediapipe==0.10.14
protobuf==4.25.3

# GUI / automation
PyAutoGUI==0.9.54
pygetwindow==0.0.9
pymsgbox==1.0.9
pyscreeze==1.0.1
pytweening==1.2.0
pyperclip==1.9.0
PyRect==0.2.0
MouseInfo==0.1.3

# Dev / notebooks (optional)
ipython==9.4.0
ipykernel==6.30.1
jupyter_client==8.6.3
jupyter_core==5.8.1

